# projects/party-in-a-box/docs/api/resident.openapi.yaml
openapi: 3.1.0
info:
  title: Party in a Box — Resident API
  version: 0.1.0
  summary: Public and resident (organizer) endpoints behind the Resident screens.
  description: |
    Contract for the Resident screens on the Party in a Box Screens canvas
    (Info, New event, Petition, My events) and the public petition sign page.
    Derived from `docs/backend-spec.md` §2, §3.1–3.2, §4 and §7. Vendor and Village
    endpoints are out of scope for this file.

    Conventions (spec §3):
    - Dates are local `YYYY-MM-DD` strings in America/Chicago (never UTC timestamps).
      Timestamps are RFC 3339 UTC.
    - Every mutation accepts an `Idempotency-Key` header.
    - Errors use the `Error` shape: `{code, message, reasons[]}`.
    - The browser may preview rule results with the shared `rules` library, but the
      server's results are authoritative.

    Privacy (spec §7): public petition responses never include signer names. Only the
    organizer of a request sees its signer list. No endpoint here exposes another
    resident's contact details.
servers:
  - url: https://api.example.org/v1
    description: Placeholder until hosting is chosen
...
paths:
...
  /threads/{threadId}/messages:
    get:
      tags: [Messages]
      operationId: listMessages
      summary: Messages in a thread the caller belongs to
      security: [{ residentAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ThreadId'
      responses:
        '200':
          description: Messages, oldest first
          content:
            application/json:
              schema:
                type: object
                required: [thread, messages]
                properties:
                  thread: { $ref: '#/components/schemas/Thread' }
                  messages:
                    type: array
                    items: { $ref: '#/components/schemas/Message' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Messages]
      operationId: postMessage
      summary: Send a message
      security: [{ residentAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ThreadId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string, minLength: 1, maxLength: 2000 }
      responses:
        '201':
          description: Sent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Message' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Unprocessable' }
components:
  securitySchemes:
    residentAuth:
      type: http
      scheme: bearer
      description: Session token from passwordless email sign-in (magic link). Issued by the auth provider; not part of this contract.
  parameters:
...
    ThreadId:
      name: threadId
      in: path
      required: true
      schema: { type: string }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, maxLength: 64 }
...
  responses:
    Unauthorized:
      description: Not signed in
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Not the organizer of this request (or not in this thread)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Unprocessable:
      description: Validation or rule check failed; `reasons` lists each one
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [code, message, reasons]
      properties:
        code: { type: string, example: submit_blocked }
        message: { type: string }
        reasons:
          type: array
          items: { type: string }
...
    Thread:
      type: object
      required: [id, kind, request_id]
      properties:
        id: { type: string }
        kind: { type: string, enum: [request, job], description: '`request` = organizer↔Village; `job` = organizer↔vendor.' }
        request_id: { type: string }
        match_id: { type: ['string', 'null'] }
    Message:
      type: object
      required: [id, thread_id, author_role, body, created_at]
      properties:
        id: { type: string }
        thread_id: { type: string }
        author_role: { type: string, enum: [resident, vendor, village] }
        author_display_name: { type: string }
        body: { type: string }
        created_at: { type: string, format: date-time }